requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
crispy-bootstrap5==0.7
Django==4.2.3
django-crispy-forms==2.0
django-filter==23.2
Pillow==10.0.0
psycopg2-binary==2.9.6
sqlparse==0.4.4
typing-extensions==4.7.1
